% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_figure.R
\name{add_figure}
\alias{add_figure}
\title{Read, render, and use a figure from a specific directory.
This function concatinates input of a specified figure with
the caption, alternative caption for accessibility, and
reference label.}
\usage{
add_figure(
  filein,
  caption = "Add figure caption",
  alt_caption = NULL,
  label = "default",
  width = 100,
  height = 100
)
}
\arguments{
\item{filein}{location of the figure to be added (example: "C:\My figure directory\plot.png")}

\item{caption}{figure caption to add below the figure in the document}

\item{alt_caption}{Default NULL. Alternative figure caption for accessibility.
If not specified the caption will be used.}

\item{label}{figure reference for citation of figure in the document}

\item{width}{figure width in terms of percentage of size. Default 100}

\item{height}{figure height in terms of perccentage of size. Default 100}
}
\value{
\code{cat} is used to print output to the screen if you run this function
on its own or to a resulting rendered file if called within an .Rmd file,
where the latter is more likely.
}
\description{
Read, render, and use a figure from a specific directory.
This function concatinates input of a specified figure with
the caption, alternative caption for accessibility, and
reference label.
}
\examples{
\dontrun{

# See below for how to include this function in your .Rmd file.

```{r, results = 'asis'}
add_figure(filein = file.path("My figure directory", "plots", "ts7_Spawning_output.png")
			  caption = "Spawning output timeseries.", 
		  alt_caption = NULL,
			  label = "ssb",
		  width = 100, height = 100)
```
}

}
\author{
Chantel R. Wetzel
}
