% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fullstop.R
\name{add_fullstop}
\alias{add_fullstop}
\title{Add a full stop to the end of a sentence}
\usage{
add_fullstop(text, punctuation = ".")
}
\arguments{
\item{text}{A text string with or without punctuation at the end.}

\item{punctuation}{The desired punctuation for the end of your string.}
}
\value{
A character string with a punctuation at the end of the string.
}
\description{
Add a full stop to the end of a sentence
}
\examples{
test <- add_fullstop("I ate an apple today")
test <- add_fullstop(c("I ate an apple today", "Example with exclamation!"))

}
\author{
Kelli F. Johnson
}
