% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors.R
\docType{data}
\name{authors}
\alias{authors}
\title{A data frame of authors and their affiliations}
\description{
A data frame of potential stock assessment authors and their
affiliations.
}
\details{
There are multiple ways to add your name to the list of authors.
You could fork the package,
add your name to the csv file \code{authors.csv} which is located in the \code{data-raw} directory, and
submit a pull request.
You could also just email the package maintainer your name, affiliation, and address.
Either option is great.
Just remember to follow the steps again when any of your information changes.
}
\keyword{data}
