% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile_changebookname}
\alias{compile_changebookname}
\title{Change bookname in the _bookdown.yml file}
\usage{
compile_changebookname(bookname)
}
\arguments{
\item{bookname}{A character value that will be used to name the output from
\link[bookdown:render_book]{bookdown::render_book}, which is normally called \verb{_main[ext]}.
This value will overwrite whatever you have in \verb{_bookdown.yml} for \code{bookname}.}
}
\description{
Change the name of the bookdown yaml file that sets up the call
to \link[bookdown:render_book]{bookdown::render_book} as suggested on
\href{https://community.rstudio.com/t/is-it-possible-to-use-variables-set-in-index-rmd-in-the-other-yaml-files/78483/5}{this Rstudio community thread}.
}
\seealso{
Other compile: 
\code{\link{compile_internal}()},
\code{\link{compile}()}
}
\author{
Kelli F. Johnson
}
\concept{compile}
