% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile_internal}
\alias{compile_internal}
\title{A non-vectorized version of \link{compile}}
\usage{
compile_internal(dir, time = 10, ...)
}
\arguments{
\item{dir}{A \emph{SINGLE} directory where the draft files are located.}

\item{time}{The number of seconds you will have to close the pdf
if it is found to be open.
The integer value is passed to \link[base:Sys.sleep]{base::Sys.sleep}.}

\item{...}{Arguments passed to \link[bookdown:render_book]{bookdown::render_book}}
}
\description{
A non-vectorized version of \link{compile} that acts as internal
code for the wrapper \link{compile}. Users typically will not interact
with this function, but the code is exported for easier viewing.
}
\seealso{
Other compile: 
\code{\link{compile_changebookname}()},
\code{\link{compile}()}
}
\author{
Kelli F. Johnson
}
\concept{compile}
