% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doit.R
\name{doit}
\alias{doit}
\title{Internal Code to Build, Load, and Render the sa Template}
\usage{
doit(
  dirgit = "c:/stockAssessment/ss/sa4ss",
  authors = c("Kelli F. Johnson", "Chantel R. Wetzel", "Joseph J. Bizzarro")
)
}
\arguments{
\item{dirgit}{A file path to where the sa4ss repository is cloned.}

\item{authors}{A vector of author names. For example,
\code{"John R. Doe"} without first name first. The middle
initial or name is optional. Entries must exactly match
those used in the \code{data-raw\authors.csv} file.
If you need additional authors added to this file
please contact the package maintainers or add them to the
\code{.csv} file and run the R script also found in that
folder prior to submitting a pull request. Following this
workflow ensures that the author's address is also added.}
}
\description{
A non-exported function used by Kelli F. Johnson to
build and load \pkg{sa4ss} and copy and render the template
for a stock assessment document.
}
\author{
Kelli Faye Johnson
}
