% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draft}
\alias{draft}
\title{Create a \pkg{sa4ss} Draft Containing Directories and Files}
\usage{
draft(type = c("ss"), create_dir = FALSE, edit = FALSE, ...)
}
\arguments{
\item{type}{The name of the template you want to copy from those available
within \pkg{sa4ss}. See the function call for available options, where the
first listed will be used for the default.}

\item{create_dir}{code{FALSE} uses the current working directory for the template.
\code{TRUE} will create a new directory and \code{"default"}, which is the
default of \code{\link[rmarkdown]{draft}()}, leaves it up to the package to
determine.}

\item{edit}{A logical value, with a default of \code{FALSE}, specifying if
you want the main file opened for editing by default. The software chosen
by R is normally not ideal and the file isn't one normally edited by users,
which is why \pkg{sa4ss} doesn't use the default value of
\code{\link[rmarkdown]{draft}}.}

\item{...}{Additional arguments that you wish to pass to
\code{\link[rmarkdown]{draft}}. See \code{args(rmarkdown::draft)}}
}
\value{
Invisibly returns the file name of the document,
which will be the main indexing file that links to the auxiliary files.
If \code{edit = TRUE}, this file will also be opened for editing.
}
\description{
Create a directory with files needed for a stock assessment
document written with \pkg{sa4ss}.
}
\details{
Code is a wrapper for \code{\link[rmarkdown]{draft}()} so users do not
have to worry about specifying all of the auxiliary arguments.
}
\examples{
\dontrun{
sa4ss::draft()
}
}
\seealso{
See \code{\link[rmarkdown]{draft}}.
}
\author{
Kelli Faye Johnson
}
