% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_table_tex.R
\name{es_table_tex}
\alias{es_table_tex}
\title{Create tex file from executive summary tables}
\usage{
es_table_tex(mod_loc, table_folder = "tables")
}
\arguments{
\item{mod_loc}{}

\item{table_folder}{}
}
\description{
Create tex file from executive summary tables
}
\author{
Chantel Wetzel
}
