% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_affiliation.R
\name{get_affiliation}
\alias{get_affiliation}
\title{Get Affiliations for Authors}
\usage{
get_affiliation(authors)
}
\arguments{
\item{authors}{A vector of author names. For example,
\code{"John R. Doe"} without first name first. The middle
initial or name is optional.}
}
\description{
Get addresses for each author using a look-up function, where
addresses are stored internally.
}
\examples{
get_affiliation("Kelli F. Johnson")
get_affiliation(c("Kelli F. Johnson", "Chantel R. Wetzel"))
# Returns an NA for the unknown name
testthat::expect_true(is.na(get_affiliation("Kelli Johnson")))
}
\author{
Kelli Faye Johnson
}
