% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_avail.R
\name{get_avail}
\alias{get_avail}
\title{Get Available \code{Rmd} Options For A Given Section}
\usage{
get_avail(
  precursor,
  dir = system.file("rmarkdown", "templates", "sa", package = "sa4ss")
)
}
\arguments{
\item{precursor}{A character string that starts with a two-digit number
and is followed by a single alpha character that specifies the type.
Specifically, everything before the \code{"-"} in the \code{Rmd} files
that are available in the package.}

\item{dir}{A file path to the directory of interest. It can be relative
or absolute. The default value looks to see the available options in the
package folder, but you can change this to be the folder for your stock
assessment, which would be helpful if you have added or deleted certain
files from the default template.}
}
\value{
Returns a vector of character values. The resulting vector is most
likely used by other functions in \code{sa4ss}.
}
\description{
Get available options in the \code{sa4ss} package for a given section.
By section, we are referring to a section in the document such as the
section on fishery-independent information, which is denoted in the
files using \code{s} and is part of the Data section, i.e., section 03.
See the example for how to get a vector of available survey types.
}
\examples{
get_avail(precursor = "11s")
}
\author{
Kelli Faye Johnson
}
