% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plotinfo.R
\name{get_plotinfo}
\alias{get_plotinfo}
\title{The following script will look for the CSV file plotInfoTable created by r4ss
in order to automate some of the plots.  The plotInfoTable file is added to the
folder each time SS_plots is run (including when you update any of the plots)
The most recent version of plotInfoTable with the complete list of plots will
be used.}
\usage{
get_plotinfo(mod_loc, plot_folder = "plots")
}
\arguments{
\item{mod_loc}{directory of the model}

\item{plot_folder}{folder name of the r4ss plots}
}
\description{
The following script will look for the CSV file plotInfoTable created by r4ss
in order to automate some of the plots.  The plotInfoTable file is added to the
folder each time SS_plots is run (including when you update any of the plots)
The most recent version of plotInfoTable with the complete list of plots will
be used.
}
\author{
Melissa Monk & Chantel Wetzel
}
