% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_.R
\name{kable_captioncontinue}
\alias{kable_captioncontinue}
\title{Create a caption for a longtable that is continued}
\usage{
kable_captioncontinue(caption = "", max = 8)
}
\arguments{
\item{caption}{The caption you want on the subsequent pages.
Typically, you will want more than the default argument of nothing,
which leads to \verb{... continued.} because this text is pasted onto
whatever you pass to this argument.}

\item{max}{The maximum number of words you want to pull from your
caption that you supplied.
The default is eight words because this typically leads to just one line.}
}
\description{
Create a caption that can be used on the repeated pages of a longtable
that spans multiple pages.
The value returned from this function should be used in
\link[kableExtra:kable_styling]{kableExtra::kable_styling} or passed to \link{kable_styling_sa4ss}.
}
\author{
Kelli F. Johnson
}
