% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_.R
\name{kable_styling_sa4ss}
\alias{kable_styling_sa4ss}
\title{Formatting for a long table with repeated headers}
\usage{
kable_styling_sa4ss(ktable, caption)
}
\arguments{
\item{ktable}{A table from \link[kableExtra:kbl]{kableExtra::kbl}.}

\item{caption}{The caption you want to be repeated on the continuing pages
of a longtable that spans more than one page.}
}
\description{
Formatting for a long table with repeated headers
}
\details{
repeat_header_continued:
The argument \code{repeat_header_continued},
which is passed to \link[kableExtra:kable_styling]{kableExtra::kable_styling},
creates a horizontal line at the bottom of your long table on every page.
Whereas, when this is \code{FALSE}, which is the default,
then the bottom of a long table is open insinuating that the table continues.
You can also add text to this argument that will be displayed below the
horizontal line.
}
\author{
Kelli F. Johnson
}
