% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_model.R
\name{read_model}
\alias{read_model}
\title{Create a \pkg{sa4ss} an RData object}
\usage{
read_model(
  mod_loc,
  plotfolder = "plots",
  printstats = FALSE,
  fecund_mult = "million eggs",
  create_plots = TRUE,
  png = TRUE,
  html = TRUE,
  datplot = TRUE,
  fleetnames = NULL,
  forecastplot = TRUE,
  maxrows = 4,
  maxcols = 4,
  bub_scale = 6,
  create_tables = TRUE,
  ci_value = 0.95,
  es_only = FALSE,
  tables = c("a", "b", "c", "d", "e", "f", "g", "h", "i", "catch", "timeseries",
    "numbers"),
  divide_by_2 = FALSE,
  adopted_ofl = NULL,
  adopted_abc = NULL,
  adopted_acl = NULL,
  forecast_ofl = NULL,
  forecast_abc = NULL,
  ...
)
}
\arguments{
\item{mod_loc}{A file path to the directory that houses the
Stock Synthesis model that will be used as the base model in this document.}

\item{plotfolder}{location of the r4ss figure directory}

\item{printstats}{Input to \code{\link[r4ss]{SS_output}} to prevent output to the screen}

\item{fecund_mult}{User input to define the multiplier for fecundity in terms of eggs if
fecundity is defined in terms of numbers within SS.  The default is millions of eggs but
there may be instances where the multiplier may be lower or higher, billions or thousands,
which can be revised by the user. If the fecundity units in SS are set as biomass then the
fecundity units (fecund_unit) will be set as metric tons.}

\item{create_plots}{TRUE//FALSE to specify whether to create model plots using r4ss}

\item{png}{TRUE create png plot files}

\item{html}{create html files}

\item{datplot}{create the data plots using \code{\link[r4ss]{SS_plots}}}

\item{fleetnames}{A vector of user-defined fleet names.
If input left as the default value of \code{NULL}, then the model-object fleet
names will be used.}

\item{forecastplot}{Add forecast years to figure plost}

\item{maxrows}{Number of rows for plots. Default set to 4.}

\item{maxcols}{Number of columns for plots. Default set to 4.}

\item{bub_scale}{Bubble scale size to use for plotting.}

\item{create_tables}{TRUE/FALSE to run \code{\link[r4ss]{SSexecutivesummary}} tables}

\item{ci_value}{To calculate confidence intervals, default is set at 0.95}

\item{es_only}{TRUE/FALSE switch to produce only the executive summary tables
will be produced, default is FALSE which will return all executive summary
tables, historical catches, and numbers-at-ages}

\item{tables}{Which tables to produce (default is everything). Note: some
tables depend on calculations related to previous tables, so will fail
if requested on their own (e.g., Table 'f' can't be created
without also creating Table 'a')}

\item{divide_by_2}{This will allow the user to calculate single sex values
based on the new sex specification (-1) in SS for single sex models. Default value is FALSE.
TRUE will divide by 2.}

\item{adopted_ofl}{Vector of adopted ofl values to be printed in the mangagement performance
table. This should be a vector of 10 values.}

\item{adopted_abc}{Vector of adopted abc values to be printed in the mangagement performance
table. This should be a vector of 10 values.}

\item{adopted_acl}{Vector of adopted acl values to be printed in the mangagement performance
table. This should be a vector of 10 values.}

\item{forecast_ofl}{Optional input vector for management adopted OFL values for table g. These values
will be overwrite the OFL values in the projection table, rather than the model estimated
OFL values. Example input: c(1500, 1300)}

\item{forecast_abc}{Optional input vector for management adopted ABC values for table g. These values
will be overwrite the ABC values in the projection table, rather than the model estimated
ABC values. Example input: c(1500, 1300)}

\item{...}{Additional input arguments passed to \code{\link[r4ss]{SS_output}} for reading
output files from Stock Synthesis into a list.}
}
\description{
Read an SS model output using \pkg{r4ss} to create a RData object with defined quantities
to be used when creating an assessment document using sa4ss
document written with \pkg{sa4ss}.
}
\details{
Read an SS model output using \pkg{r4ss} to create a RData object with defined quantities
to be used when creating an assessment document using sa4ss
document written with \pkg{sa4ss}.
}
\examples{
\dontrun{
simplemod_loc <- tail(dir(pattern = "simple",
  system.file("extdata", package = "r4ss"),
  full.names = TRUE), 1)
sa4ss::read_model(mod_loc = simplemod_loc,
  fecund_mult = "billion eggs")
}
}
\author{
Chantel Wetzel
}
