% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_figure.R
\name{report_ref_label}
\alias{report_ref_label}
\title{Format a figure or table label for use in a report}
\usage{
report_ref_label(x)
}
\arguments{
\item{x}{A vector of strings. The strings should be valid figure or table
labels. For example \code{x = c("fig:apple", "fig:banana")}.}
}
\description{
Format a text string that was used to label a figure or a table such that
you can reference that figure or table elsewhere in a report. For example the
label might be \code{fig:my-best-figure} and in the text you might want \verb{Look at the beautiful pie chart in my best figure ever (Figure \\ref\{fig:mybest-figure\})}.
}
\examples{
# Generate some figure references
report_ref_label("fig:apple")
report_ref_label(c("fig:apple", "fig:banana"))
report_ref_label(c("fig:apple", "fig:banana", "fig:carrot"))
# Works for tables too
report_ref_label("tab:apple")
}
\seealso{
\itemize{
\item \code{\link[=add_figure]{add_figure()}}
}
}
\author{
Kelli F. Johnson
}
