% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa4ss-package.R
\docType{package}
\name{sa4ss-package}
\alias{sa4ss}
\alias{sa4ss-package}
\title{sa4ss: Write Stock Assessments for Stock Synthesis}
\description{
Create files and the structure necessary to facilitate rapid and reproducible creation of stock assessment documents from model output generated by Stock Synthesis (SS3). The included components and structure are specific to requests from the Pacific Fishery Management Council for U.S. west coast groundfish.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pfmc-assessments/sa4ss}
  \item Report bugs at \url{https://github.com/pfmc-assessments/sa4ss/issues}
}

}
\author{
\strong{Maintainer}: Kelli F. Johnson \email{kelli.johnson@noaa.gov} (\href{https://orcid.org/0000-0002-5149-451X}{ORCID})

Authors:
\itemize{
  \item Chantel R. Wetzel \email{chantel.wetzel@noaa.gov} (\href{https://orcid.org/0000-0002-7573-8240}{ORCID})
}

}
\keyword{internal}
