% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{session_test}
\alias{session_test}
\title{Test your installation of sa4ss}
\usage{
session_test()
}
\value{
A list object is returned with two entries.
The first entry, \code{session_info},
contains stored information about your work session.
The typical output from \link{Sys.getenv} is augmented with information pertinent
to the sa4ss package and compiling accessible pdf documents.
The second entry, \code{tried}, is the output from \link{tryCatch} when attempting
to compile the .Rmd files into a pdf using \link[bookdown:render_book]{bookdown::render_book}.
If the function fails to produce a successful pdf, then a
\link[utils:bug.report]{utils::bug.report} is generated and users are sent to github issues,
where they can fill out an issue based on the installation issue template.
}
\description{
Test the installation of sa4ss to build the template pdf.
}
\details{
sa4ss relies on a lot of back-end functionality that typical R packages
do not require. Ensuring that these are installed properly is facilitated
when you use Rstudio because the developers of Rstudio have put in a lot of
work towards making it easy for users to compile documents. sa4ss also relies
on \code{sed} which is available in the tools provided by R, i.e., rtools40, but
not always available in your path.

The premise of this function is to store information if the pdf build is not
successful that is of use to the developers. If you provide the generic output
that is returned to the developers, then they will be able to help you and
future individuals like yourself work through installation problems.
So, thank you for being willing to help.
}
\examples{
\dontrun{
localtest <- session_test()
}
}
\author{
Kelli F. Johnson
}
