% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_format.R
\name{table_format}
\alias{table_format}
\title{Custom table for sa4ss}
\usage{
table_format(
  x,
  format = "latex",
  caption = "Add Caption",
  label = NULL,
  digits = getOption("digits"),
  booktabs = TRUE,
  col_names = NULL,
  linesep = "",
  longtable = TRUE,
  font_size = 10,
  align = "c",
  col_names_align = "c",
  hold_position = TRUE,
  escape = TRUE,
  bold_header = FALSE,
  landscape = FALSE,
  repeat_header = FALSE,
  header_grouping = NULL,
  format.args = NULL,
  custom_width = FALSE,
  col_to_adjust = NULL,
  width = NULL,
  create_png = lifecycle::deprecated(),
  ...
)
}
\arguments{
\item{x}{An R object, typically a matrix or data frame.}

\item{format}{As defined by \code{\link[kableExtra:kbl]{kableExtra::kbl()}}. Default for sa4ss is 'latex'.}

\item{caption}{As defined by \code{\link[kableExtra:kbl]{kableExtra::kbl()}}.}

\item{label}{As defined by \code{\link[kableExtra:kbl]{kableExtra::kbl()}}.}

\item{digits}{As defined by \code{\link[kableExtra:kbl]{kableExtra::kbl()}}.}

\item{booktabs}{As defined by \code{\link[kableExtra:kbl]{kableExtra::kbl()}}. Logical.}

\item{col_names}{Names for the columns to show on table.}

\item{linesep}{As defined by \code{\link[kableExtra:kbl]{kableExtra::kbl()}}.}

\item{longtable}{As defined by \code{\link[kableExtra:kbl]{kableExtra::kbl()}}.}

\item{font_size}{Font size in pts. If NULL, document font size is used.}

\item{align}{As defined by \code{\link[kableExtra:kbl]{kableExtra::kbl()}}.}

\item{col_names_align}{As defined in \code{\link[kableExtra:linebreak]{kableExtra::linebreak()}}.}

\item{hold_position}{As defined in \code{\link[kableExtra:kable_styling]{kableExtra::kable_styling()}}. Logical.}

\item{escape}{As defined by \code{\link[kableExtra:kable_styling]{kableExtra::kable_styling()}}.}

\item{bold_header}{Make headers bold. Logical}

\item{landscape}{Make this table in landscape orientation?}

\item{repeat_header}{If landscape, repeat the header on subsequent pages?}

\item{header_grouping}{As defined by \code{\link[kableExtra:kable_styling]{kableExtra::kable_styling()}}}

\item{format.args}{As defined by \code{\link[kableExtra:kbl]{kableExtra::kbl()}}.}

\item{custom_width}{Logical. Allow for custom column widths}

\item{col_to_adjust}{Vector of columns to adjust width. Only used if custom_width = TRUE.}

\item{width}{Vector or single value of column widths (i.e. c('2cm', '2cm')) for the columns defined
in the col_to_adjust.}

\item{create_png}{DEPRECATED. Logical. If set to true tables will be created
as png objects in the doc.}

\item{...}{Extra arguments supplied to \code{\link[kableExtra]{kbl}}.}
}
\description{
Custom table for sa4ss
}
\examples{
table_format(head(iris))
}
