% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{techreport_pdf}
\alias{techreport_pdf}
\title{Technical report output formatting}
\usage{
techreport_pdf(
  latex_engine = c("lualatex", "pdflatex"),
  pandoc_args = c("--top-level-division=section", "--wrap=none",
    "--default-image-extension=png", "--lua-filter=tagged-filter.lua"),
  ...
)
}
\arguments{
\item{latex_engine}{The desired latex engine, where luaLaTex is the
default because it leads to a 508 compliant document. This argument
is passed to \code{\link[bookdown]{pdf_book}}. Only a single value
can be passed, but all options are listed below.}

\item{pandoc_args}{Arguments for pandoc. This argument
is passed to \code{\link[bookdown]{pdf_book}}.}

\item{...}{Additional arguments passed to \code{\link[bookdown]{pdf_book}}.}
}
\description{
Format for creating a technical report with the \code{.pdf} extension.
This format will be used to generate markdown from R Markdown and a
resulting pdf will be saved to the disk.
}
