% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_authors.R
\name{write_authors}
\alias{write_authors}
\title{Compile Authors Names in YAML Form}
\usage{
write_authors(authors, fileout = "00authors.Rmd")
}
\arguments{
\item{authors}{A vector of author names. For example,
\code{"John R. Doe"} without first name first. The middle
initial or name is optional.}

\item{fileout}{A file path for the file that will be saved to the disk.
Typically, full or relative paths are allowed.}
}
\description{
Write a file to the disk with all authors and their affiliations
using the yaml form for inclusion in the stock assessment.
}
\author{
Kelli Faye Johnson
}
