% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_summary.R
\name{write_summary}
\alias{write_summary}
\title{Write Summary Files}
\usage{
write_summary(type = c("executive", "onepage"), fileout = "01summaries.Rmd")
}
\arguments{
\item{type}{A vector of the type of summaries you want included in the stock
assessment document. Current options are listed in the default call to
\code{write_summary()}.}

\item{fileout}{A file path for the file that will be saved to the disk.
Typically, full or relative paths are allowed.}
}
\description{
Write a file to the disk with the title using the yaml form for inclusion in
the stock assessment.
}
\author{
Kelli F. Johnson
}
