% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_title.R
\name{write_title}
\alias{write_title}
\title{Compile Title in YAML format}
\usage{
write_title(
  species,
  latin,
  coast,
  year = format(Sys.Date(), "\%Y"),
  fileout = "00title.Rmd"
)
}
\arguments{
\item{species}{A character string specifying the species name.}

\item{latin}{A character string specifying the latin name.}

\item{coast}{A character string specifying which part of the coast
the stock assessment includes.}

\item{year}{A numeric or character string specifying the year.
The default uses \code{Sys.Date()} to format the year into a four
digit character value.}

\item{fileout}{A file path for the file that will be saved to the disk.
Typically, full or relative paths are allowed.}
}
\description{
Write a file to the disk with the title
using the yaml form for inclusion in the stock assessment.
}
\author{
Kelli Faye Johnson
}
