% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draft}
\alias{draft}
\title{Create a \pkg{sa4ss} Draft Containing Directories and Files}
\usage{
draft(
  authors,
  species = "Species name",
  latin = "Scientific name",
  coast = "US West",
  type = c("sa"),
  create_dir = FALSE,
  edit = FALSE,
  ...
)
}
\arguments{
\item{authors}{A vector of author names. For example,
\code{"John R. Doe"} without first name first. The middle
initial or name is optional. Entries must exactly match
those used in the \code{"data-raw\authors.csv"} file.
If you need additional authors added to this file
please contact the package maintainers or add them to the
\code{.csv} file and run the R script also found in that
folder prior to submitting a pull request. Following this
workflow ensures that the author's address is also added.}

\item{species}{A vector of character strings providing species names for
all species included in the stock assessment. Typically, there will only
be one species, but this template can accommodate a species complex as well.
Each species should be capitalized as you would like to see it used in the middle
of a sentence. For example,
\code{"Pacific Hake"} is used in the stock assessment of hake;
not \code{"pacific hake"} because Pacific Hake is the preferred name, not
pacific hake. The text can be called in the document using spp for the
form to be used in the middle of a sentence and Spp when placed at the
beginning of a sentence.}

\item{latin}{A character string specifying the Latin name without
italic formatting. Please capitalize as you want it to be used in
the document though because case will not be altered.}

\item{coast}{A character string specifying which part of the coast
the stock assessment includes. For example, \code{"US West"}.
Do not include \code{"coast"} because it will be added internally.}

\item{type}{The name of the template you want to copy from those available
within \pkg{sa4ss}. See the function call for available options, where the
first listed will be used for the default.}

\item{create_dir}{code{FALSE} uses the current working directory for the template.
\code{TRUE} will create a new directory and \code{"default"}, which is the
default of \code{\link[rmarkdown]{draft}()}, leaves it up to the package to
determine.}

\item{edit}{A logical value, with a default of \code{FALSE}, specifying if
you want the main file opened for editing by default. The software chosen
by R is normally not ideal and the file isn't one normally edited by users,
which is why \pkg{sa4ss} doesn't use the default value of
\code{\link[rmarkdown]{draft}}.}

\item{...}{Additional arguments that you wish to pass to
\code{\link[rmarkdown]{draft}}. See \code{args(rmarkdown::draft)}}
}
\value{
Invisibly returns the file name of the document,
which will be the main indexing file that links to the auxiliary files.
If \code{edit = TRUE}, this file will also be opened for editing.
}
\description{
Create a directory with files needed for a stock assessment
document written with \pkg{sa4ss}.
}
\details{
Along with specifying some initial inputs, such as
authors and species, this code is a wrapper for \code{\link[rmarkdown]{draft}()}.
Users should focus on the input arguments that come before \code{type}
because these will be specific to your stock assessment and are used to
set up the initial file structure.
}
\examples{
\dontrun{
sa4ss::draft()
}
}
\seealso{
See \code{\link[rmarkdown]{draft}}.
}
\author{
Kelli Faye Johnson
}
