% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_table_tex.R
\name{es_table_tex}
\alias{es_table_tex}
\title{Create tex file from executive summary tables}
\usage{
es_table_tex(
  dir,
  table_folder = NULL,
  save_loc = NULL,
  csv_name = "table_labels.csv"
)
}
\arguments{
\item{dir}{The directory to look for the csv file that has a list of all
csv files to create tex files for. The csv file format should match
the table_labels.csv created by r4ss function SSexecutivesummary. The columns
names in this folder are caption, altcaption, label, filename, and loc (optional).
The loc column is a directory location to find the file (filename) that
a tex file should be created for. The \code{dir} can be set to the
\code{mod_loc} where the function will automatically look for them inside a
table folder in that location.}

\item{table_folder}{The relative path for the table directory, where it
must be relative to \code{dir} because it will be appended to this
argument using \code{file.path}. An alternative is to supply the full
file path in dir and use \code{table_folder = ""}.}

\item{save_loc}{optional input that requires a full path which will
allow users to save the tex file to a specific location (e.g. inside doc folder)
other than the \code{file.path(dir, table_folder)}. Default NULL}

\item{csv_name}{CSV file name to create tex table scripts for. This file should have
the following columns: caption, altcaption, label, filenem, loc (optional) where caption will
be the table caption, altcapation is the accessibility text for the table (can be
NA if the caption should be used), label is the text to reference the table, filename
is the file name of the csv file to be read, and loc (optional) is the location to file the
csv file indicated by the filename.}
}
\description{
Create tex file from executive summary tables
}
\examples{
\dontrun{
#example of creating tex files for those created by the 
 # r4ss function SSexecutivesummary
	es_table_tex(dir = mod_loc)
#example of using a user created csv file:
es_table_tex(dir = "C:/models/table_to_add_to_doc",
			 save_loc = "C:/model/tex_tables",
			 csv_name = "non_es_document_tables.csv")
}			
}
\author{
Chantel Wetzel
}
