% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_affiliation.R
\name{get_affiliation}
\alias{get_affiliation}
\title{Get Affiliations for Authors}
\usage{
get_affiliation(authors)
}
\arguments{
\item{authors}{A vector of author names. For example,
\code{"John R. Doe"} without first name first. The middle
initial or name is optional. Entries must exactly match
those used in the \code{"data-raw\authors.csv"} file.
If you need additional authors added to this file
please contact the package maintainers or add them to the
\code{.csv} file and run the R script also found in that
folder prior to submitting a pull request. Following this
workflow ensures that the author's address is also added.}
}
\value{
A vector the same length as the input author argument is returned,
providing the matching address for each entry as a text string.
}
\description{
Get addresses for each author using a look-up function, where
addresses are stored internally.
}
\examples{
get_affiliation("Kelli F. Johnson")
get_affiliation(c("Kelli F. Johnson", "Chantel R. Wetzel"))
# Returns an NA for the unknown name
testthat::expect_true(is.na(get_affiliation("Kelli Johnson")))
}
\author{
Kelli F. Johnson
}
