% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_authors.R
\name{write_authors}
\alias{write_authors}
\title{Compile Authors Names in YAML Form}
\usage{
write_authors(authors, fileout = "00authors.Rmd")
}
\arguments{
\item{authors}{A vector of author names. For example,
\code{"John R. Doe"} without first name first. The middle
initial or name is optional. Entries must exactly match
those used in the \code{"data-raw\authors.csv"} file.
If you need additional authors added to this file
please contact the package maintainers or add them to the
\code{.csv} file and run the R script also found in that
folder prior to submitting a pull request. Following this
workflow ensures that the author's address is also added.}

\item{fileout}{A file path for the file that will be saved to the disk.
Typically, full or relative paths are allowed.}
}
\description{
Write a file to the disk with all authors and their affiliations
using the yaml form for inclusion in the stock assessment.
}
\examples{
# An example with a standard first name middle initial full stop and last name
# then a first initial full stop middle initial full stop and last name and
# last a first name and last name with no middle initial.
eg <- utils::capture.output(
  write_authors(
    c("Kelli F. Johnson", "E. J. Dick", "Qi Lee"),
    fileout = ""
  )
)
\dontrun{
print(eg)
}
}
\author{
Kelli Faye Johnson
}
